Split tag logic to it's own module
